Second try to get an expect module. This time, it uses pexpect (http://pexpect.sourceforge.net/), so this adds another dependency, but this is just for the spexpect module.
Needed to runtime-patch the pexpect.spawn class though, since it uses
os.write and os.read, and these functions don't play well with paramiko
channels.

I will try to make a generic patch and submit it to the pexpect maintainer.

The spexpect module is still very rough, since it doesn't handle errors and
timeouts. The script language is not yet well defined, so this is just a
toy right now.