Big refactoring of the console.
Now the console is able to 'talk' to both the underlying server (console is
forked from the server to be able to use readline), and the main daemon
process.
This allows to communicate with sibling processes to watch, message or kill
them.

The commands available on the console are also potentially available to the
command line, it's just a matter of adding new options to the command line
parser (easily done via plugins).